% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Build_AUCell_Rankings}
\alias{Build_AUCell_Rankings}
\title{Building AUCell gene rankings}
\usage{
Build_AUCell_Rankings(
  object,
  AUCell_buildRankings.file = "",
  nCores = 1,
  plotStats = FALSE
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_buildRankings.file}{The output file of AUCell rankings.}

\item{nCores}{The number of cores for running AUCell. Default 1}

\item{plotStats}{is logical. If TRUE, the AUCell stats will be displayed.}
}
\description{
Building AUCell gene rankings
}
