% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Build_AUCell_Rankings_Fast}
\alias{Build_AUCell_Rankings_Fast}
\title{Fast building AUCell gene rankings}
\usage{
Build_AUCell_Rankings_Fast(
  object,
  AUCell_buildRankings.file = "",
  nCores = 1,
  block.cell.size = 5000,
  plotStats = FALSE
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_buildRankings.file}{The output file of AUCell rankings.}

\item{nCores}{The number of cores for running AUCell. Default 1}

\item{block.cell.size}{The number of genes in each block of AUCell processing. Default 5000}

\item{plotStats}{is logical. If TRUE, the AUCell stats will be displayed.}
}
\description{
Fast building AUCell gene rankings
}
