% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{clustering_KMeansOnKNN_Graph}
\alias{clustering_KMeansOnKNN_Graph}
\title{Run K-Means clustering on the KNN-graph}
\usage{
clustering_KMeansOnKNN_Graph(object, k = 10, iter.max = 100)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{k}{The number of k-means clusters.}

\item{iter.max}{The maximum number of k-mean's iterations.}
}
\description{
Run K-Means clustering on the KNN-graph
}
