% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{DoubletDetection_with_scrublet}
\alias{DoubletDetection_with_scrublet}
\title{DoubletDetection_with_scrublet}
\usage{
DoubletDetection_with_scrublet(
  object,
  expected.doublet.rate = 0.03,
  seed = 2021L,
  PCs = 30L,
  min.cells = 10,
  min_gene_variablity = 85
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{expected.doublet.rate}{The expected fraction of transcriptomes that are doublets, typically 0.05-0.1.}

\item{seed}{The seed number}

\item{PCs}{The number of input PCA components}

\item{min.cells}{Used for gene filtering prior to PCA. Genes expressed at fewer than min_counts, default=3, in fewer than min_cells are excluded.}

\item{min_gene_variablity}{Used for gene filtering prior to PCA. Keep the most highly variable genes for the analysis.}
}
\description{
DoubletDetection_with_scrublet
}
