% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{export_marker_genes_to_table}
\alias{export_marker_genes_to_table}
\title{Export identified marker genes to table.}
\usage{
export_marker_genes_to_table(
  object,
  cluster.type = c("walktrap", "louvain", "kmeans", "merged_louvain", "merged_walktrap",
    "merged_kmeans"),
  n.TopGenes = 5,
  min.log2FC = 0.5,
  min.expFraction = 0.3,
  write.to.file = ""
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{n.TopGenes}{The number of top differential genes. Default 5}

\item{min.log2FC}{The minimum log2FC.}

\item{min.expFraction}{The minimum fraction of expressing cell frequency.}

\item{write.to.file}{The output file name.}
}
\description{
Export identified marker genes to table.
}
