% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{filter_cells_and_genes}
\alias{filter_cells_and_genes}
\title{Filter out cells and genes}
\usage{
filter_cells_and_genes(
  object,
  min_UMIs = 1000,
  max_UMIs = 30000,
  min_detected_genes = 1000,
  max_detected_genes = 8000,
  min_percent_mito = 0,
  max_percent_mito = 10,
  genes_with_expressing_cells = 10,
  isRemovedDoublets = TRUE
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{min_UMIs}{The minimum UMI cutoff.}

\item{max_UMIs}{The maximum UMI cutoff.}

\item{min_detected_genes}{The minimum number of genes detected per cell.}

\item{max_detected_genes}{The maximum number of genes detected per cell.}

\item{min_percent_mito}{The minimum percent of UMIs from the mitocondrial genes.}

\item{max_percent_mito}{The maximum percent of UMIs from the mitocondrial genes.}

\item{genes_with_expressing_cells}{The cutoff for genes expressing in at least a number of cells.}

\item{isRemovedDoublets}{remove doublets or not, default = TRUE. This requires doublet detection result from the function 'DoubletDetection_with_scrublet'}
}
\description{
Filter out cells and genes
}
