% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{findMarkerGenes}
\alias{findMarkerGenes}
\title{Identify marker genes for each cluster.}
\usage{
findMarkerGenes(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans", "merged_walktrap",
    "merged_louvain", "merged_kmeans"),
  min.log2FC = 0.5,
  min.expFraction = 0.3,
  limit.top.genes = 500,
  use.sampling.cells = FALSE,
  use.fraction.of.cells = 0.2,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The type of clustering method.}

\item{min.log2FC}{The minimum cutoff of the log2FC. Default 0.5}

\item{min.expFraction}{The minimum expressing cells fraction. Default 0.3}

\item{limit.top.genes}{The limited number of differential genes to be shown. Default 500}

\item{use.sampling.cells}{is logical. If TRUE, the downsample cells will be processed.}

\item{use.fraction.of.cells}{The fraction of downsample. Default 0.2}

\item{n.seed}{The number of set seed.}
}
\description{
Identify marker genes for each cluster.
}
