% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_cells_annotation}
\alias{get_cells_annotation}
\title{Get cell annotation}
\usage{
get_cells_annotation(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get cell annotation
}
