% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_dfm.result}
\alias{get_dfm.result}
\title{Get diffusion map analysis result}
\usage{
get_dfm.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get diffusion map analysis result
}
