% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_dir_path_10x_matrix}
\alias{get_dir_path_10x_matrix}
\title{Get path directory of 10x data}
\usage{
get_dir_path_10x_matrix(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get path directory of 10x data
}
