% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_fa2_graph.layout}
\alias{get_fa2_graph.layout}
\title{Get FA2 graph-layout}
\usage{
get_fa2_graph.layout(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get FA2 graph-layout
}
