% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_genes_metadata}
\alias{get_genes_metadata}
\title{Get gene's metadata}
\usage{
get_genes_metadata(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get gene's metadata
}
