% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellUtils.R
\name{get_gmtGeneSets}
\alias{get_gmtGeneSets}
\title{Get gene sets from GMT file}
\usage{
get_gmtGeneSets(gmt.file)
}
\arguments{
\item{gmt.file}{The GMT file name.}
}
\value{
List of gene sets.
}
\description{
Get gene sets from GMT file
}
