% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_knn_graph.graph}
\alias{get_knn_graph.graph}
\title{Get graph from KNN-graph analysis result}
\usage{
get_knn_graph.graph(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get graph from KNN-graph analysis result
}
