% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_knn_graph.layout}
\alias{get_knn_graph.layout}
\title{Get graph-layout from KNN-graph analysis result}
\usage{
get_knn_graph.layout(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get graph-layout from KNN-graph analysis result
}
