% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_lsi.result}
\alias{get_lsi.result}
\title{Get LSI analysis result}
\usage{
get_lsi.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get LSI analysis result
}
