% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Get marker genes information}
\usage{
get_marker_genes(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get marker genes information
}
