% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_nnmf.result}
\alias{get_nnmf.result}
\title{Get NNMF analysis result}
\usage{
get_nnmf.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get NNMF analysis result
}
