% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_normalized_umi}
\alias{get_normalized_umi}
\title{Get normalized UMI count matrix}
\usage{
get_normalized_umi(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get normalized UMI count matrix
}
