% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_pca.result}
\alias{get_pca.result}
\title{Get PCA analysis result}
\usage{
get_pca.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get PCA analysis result
}
