% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_regressout_log_data}
\alias{get_regressout_log_data}
\title{Get regressed out log-scaled matrix}
\usage{
get_regressout_log_data(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get regressed out log-scaled matrix
}
