% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_sample_uniq_id}
\alias{get_sample_uniq_id}
\title{Get unique sample ID}
\usage{
get_sample_uniq_id(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get unique sample ID
}
