% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_tsne.result}
\alias{get_tsne.result}
\title{Get tSNE analysis result}
\usage{
get_tsne.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get tSNE analysis result
}
