% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellGenerics.R
\name{get_umap.result}
\alias{get_umap.result}
\title{Get UMAP analysis result}
\usage{
get_umap.result(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Get UMAP analysis result
}
