% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{get_variable_genes_basic}
\alias{get_variable_genes_basic}
\title{Identified variable genes by using the standard mean of gene expression and coefficient of variation}
\usage{
get_variable_genes_basic(
  object,
  mean_expr_cutoff = 0.1,
  max_of_mean_expr_cutoff = 5,
  cv.max = 3,
  cv.min = 0.5
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{mean_expr_cutoff}{The mean expression cutoff. Default 0.1}

\item{max_of_mean_expr_cutoff}{The maximum number of gene expression cutoff. This is to filter out housekeeping and ribosomal genes. Default 5}

\item{cv.max}{The maximum CV cutoff. Default 3}

\item{cv.min}{The minimum CV cutoff. Default 0.5}
}
\description{
Identified variable genes by using the standard mean of gene expression and coefficient of variation
}
