% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{getCellsFromClusters}
\alias{getCellsFromClusters}
\title{Get cell names from selected clusters.}
\usage{
getCellsFromClusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  cluster_id = c("cl1", "cl2")
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The type of clustering method.}

\item{cluster_id}{The specified cluster ID (e.g. cluster1)}
}
\value{
a vector of cell names.
}
\description{
Get cell names from selected clusters.
}
