% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{getMarkerGenesInfo}
\alias{getMarkerGenesInfo}
\title{Get marker genes information for any selected cluster.}
\usage{
getMarkerGenesInfo(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans", "merged_louvain", "merged_walktrap",
    "merged_kmeans"),
  cluster_id = "cl1"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The type of clustering method.}

\item{cluster_id}{The specified cluster ID (e.g. cluster1)}
}
\value{
a dataframe of marker information
}
\description{
Get marker genes information for any selected cluster.
}
