% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{identify_potentially_merging_clusters}
\alias{identify_potentially_merging_clusters}
\title{Identify the clusters that potenntially can be merged.}
\usage{
identify_potentially_merging_clusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  min.log2FC = 0.3,
  min.expFraction = 0.3,
  min.jaccard_similarity = 0.5
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.3}

\item{min.expFraction}{The minimum fraction of expressing cell frequency cutoff. Default 0.3}

\item{min.jaccard_similarity}{The minimum jaccard similarity.}
}
\description{
Identify the clusters that potenntially can be merged.
}
