% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{identifyDifferentialGenes_in_a_cluster_vs_the_rest_of_clusters}
\alias{identifyDifferentialGenes_in_a_cluster_vs_the_rest_of_clusters}
\title{Identify differentially expressed genes between a selected cluster vs the rest of clusters}
\usage{
identifyDifferentialGenes_in_a_cluster_vs_the_rest_of_clusters(
  object,
  cluster.id = "",
  method = c("wilcoxon"),
  cluster.type = c("louvain", "walktrap", "kmeans", "merged_walktrap", "merged_louvain",
    "merged_kmeans"),
  min.log2FC = 0.25,
  min.expFraction = 0.3,
  write.to.file = ""
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The type of clustering method.}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.25}

\item{min.expFraction}{The minimum fraction of expressing cell frequency cutoff. Default 0.3}

\item{write.to.file}{The output file name.}

\item{cluster_id}{The specified cluster ID (e.g. cluster1)}
}
\value{
a dataframe of differentially expressed genes.
}
\description{
Identify differentially expressed genes between a selected cluster vs the rest of clusters
}
