% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{identifyGSEAPrerankedGenes}
\alias{identifyGSEAPrerankedGenes}
\title{Identify GSEA preranked genes}
\usage{
identifyGSEAPrerankedGenes(
  method = c("wilcoxon"),
  objectA = objectA,
  objectB = objectB,
  cellsA = c(),
  cellsB = c(),
  write.to.file = "",
  min.log2FC = 0.1,
  min.expFraction = 0.01
)
}
\arguments{
\item{method}{The differential gene expression testing method. Default 'wilcoxon'  test}

\item{objectA}{The SingCellaR object A}

\item{objectB}{The SingCellaR object B}

\item{cellsA}{The vector of cell names from object A}

\item{cellsB}{The vector of cell names from object B}

\item{write.to.file}{The output file name.}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.1}

\item{min.expFraction}{The minimum fraction of expressing cell frequency cutoff. Default 0.01}
}
\description{
Identify GSEA preranked genes
}
