% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{identifyGSEAPrerankedGenes_for_all_clusters}
\alias{identifyGSEAPrerankedGenes_for_all_clusters}
\title{Identify GSEA's preranked genes for all identified clusters}
\usage{
identifyGSEAPrerankedGenes_for_all_clusters(
  object,
  cluster.type = c("louvain", "kmeans", "merged_walktrap", "merged_louvain",
    "merged_kmeans"),
  fishers_exact_test = 0.1,
  min.expFraction = 0.01,
  min.log2FC = 0.1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The type of clustering method.}

\item{fishers_exact_test}{The fisher's exact test cutoff p-value. Default 0.1}

\item{min.expFraction}{The minimum fraction of expressing cell frequency cutoff. Default 0.01}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.1}
}
\description{
Identify GSEA's preranked genes for all identified clusters
}
