% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{identifyGSEAPrerankedGenes_in_a_cluster_vs_the_rest_of_clusters}
\alias{identifyGSEAPrerankedGenes_in_a_cluster_vs_the_rest_of_clusters}
\title{Identify GSEA's preranked genes for a selected cluster vs the rest of clusters}
\usage{
identifyGSEAPrerankedGenes_in_a_cluster_vs_the_rest_of_clusters(
  object,
  cluster.id = "",
  cluster.type = c("louvain", "kmeans", "merged_walktrap", "merged_louvain",
    "merged_kmeans"),
  fishers_exact_test = 0.1,
  min.expFraction = 0.01,
  min.log2FC = 0.1,
  write.to.file = ""
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.id}{The specified cluster id (e.g. 'cluster1')}

\item{cluster.type}{The type of clustering method.}

\item{fishers_exact_test}{The fisher's exact test cutoff p-value. Default 0.1}

\item{min.expFraction}{The minimum fraction of expressing cell frequency cutoff. Default 0.01}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.1}

\item{write.to.file}{The output file name.}
}
\value{
a dataframe of preranked genes.
}
\description{
Identify GSEA's preranked genes for a selected cluster vs the rest of clusters
}
