% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellUtils.R
\name{listAvailableGeneSets}
\alias{listAvailableGeneSets}
\title{List of available gene sets}
\usage{
listAvailableGeneSets(gmt.file)
}
\arguments{
\item{gmt.file}{The GMT file name.}
}
\description{
List of available gene sets
}
