% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{load_gene_expression_from_a_file}
\alias{load_gene_expression_from_a_file}
\title{Load gene expression from a file}
\usage{
load_gene_expression_from_a_file(
  object,
  input_file = "",
  isTargetSeq = T,
  sep = ","
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{input_file}{The input gene expression file name.}

\item{isTargetSeq}{is the logical input. default = TRUE}

\item{sep}{The field separator character of the gene expression table.}
}
\description{
Load gene expression from a file
}
