% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{load_matrices_from_cellranger}
\alias{load_matrices_from_cellranger}
\title{Load 10X sparse data matrices provided by the Cell Ranger software}
\usage{
load_matrices_from_cellranger(
  object,
  cellranger.version = 3,
  isMultiFeatures = FALSE,
  selectedFeature = "Gene Expression"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cellranger.version}{is the version of the Cell Ranger software used to generate input matrices.}

\item{isMultiFeatures}{is logical}

\item{selectedFeature}{SingCellaR supports two features 'Gene Expression' and 'Antibody Capture'}
}
\description{
Load 10X sparse data matrices provided by the Cell Ranger software
}
