% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merged clusters}
\usage{
merge_clusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  merge_cluster_ids = c()
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{merge_cluster_ids}{The vector of cluster ids that will be merged together e.g. c('cl1:cl2','cl9:cl10').}
}
\description{
Merged clusters
}
