% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{normalize_UMIs}
\alias{normalize_UMIs}
\title{Normalisation}
\usage{
normalize_UMIs(object, scale.factor = 10000, use.scaled.factor = T)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{scale.factor}{The normalised scale factor which is usually set up at 10000.}

\item{use.scaled.factor}{default is TRUE if it is FALSE, the function will use the mean of library size as the scale factor.}
}
\description{
Normalisation
}
