% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_3D_knn_graph_label_by_a_signature_gene_set}
\alias{plot_3D_knn_graph_label_by_a_signature_gene_set}
\title{Plot KNN-Graph trajectory in 3D with the gene expression of a signature gene set}
\usage{
plot_3D_knn_graph_label_by_a_signature_gene_set(
  object,
  gene_list = c(),
  isNormalizedByHouseKeeping = F,
  vertext.size = 0.4,
  point.color1 = "black",
  point.color2 = "red"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gene_list}{The vector of gene names.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{vertext.size}{The node size. Default 0.4}

\item{point.color1}{The low color gradient of gene expression. Default black}

\item{point.color2}{The high color gradient of gene expression. Default red}
}
\description{
Plot KNN-Graph trajectory in 3D with the gene expression of a signature gene set
}
