% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_3D_knn_graph_label_by_clusters}
\alias{plot_3D_knn_graph_label_by_clusters}
\title{Plot KNN-Graph trajectory in 3D with the cluster name}
\usage{
plot_3D_knn_graph_label_by_clusters(
  object,
  show_method = c("walktrap", "louvain", "kmeans", "merged_walktrap", "merged_louvain",
    "merged_kmeans"),
  vertext.size = 0.4,
  edge.color = "gray45"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{show_method}{The vector of clustering method names.}

\item{vertext.size}{The node size. Default 0.4}

\item{edge.color}{The edge color. Default gray45}
}
\description{
Plot KNN-Graph trajectory in 3D with the cluster name
}
