% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_3D_knn_graph_label_by_gene}
\alias{plot_3D_knn_graph_label_by_gene}
\title{Plot KNN-Graph trajectory in 3D with the gene expression of selected genes}
\usage{
plot_3D_knn_graph_label_by_gene(
  object,
  show.gene = c(),
  vertext.size = 0.4,
  point.color1 = "black",
  point.color2 = "red",
  edge.color = "gray45"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{show.gene}{The vector of gene names.}

\item{vertext.size}{The node size. Default 0.4}

\item{point.color1}{The low color gradient of gene expression. Default black}

\item{point.color2}{The high color gradient of gene expression. Default red}

\item{edge.color}{The edge color. Default gray45}
}
\description{
Plot KNN-Graph trajectory in 3D with the gene expression of selected genes
}
