% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_bubble_for_genes_per_cluster}
\alias{plot_bubble_for_genes_per_cluster}
\title{Plot bubble-plot for gene expression per cluster}
\usage{
plot_bubble_for_genes_per_cluster(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  gene_list = c(),
  point.color1 = "orange",
  point.color2 = "firebrick",
  buble.scale = 6,
  gene.font.size = 12,
  axist.x.font.size = 12,
  show.percent = FALSE,
  log = T,
  font.size.in.bubble = 2
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{gene_list}{The vector of gene names.}

\item{point.color1}{The color for low expression level. Default orange}

\item{point.color2}{The color for high expression level. Default firebrick}

\item{buble.scale}{The scale size of the bubble. Default 6}

\item{gene.font.size}{The font size of gene name. Default 12}

\item{axist.x.font.size}{The label on the x-axis font size. Default 12}

\item{show.percent}{is logical. If TRUE, the percentage will be displayed.}

\item{log}{is logical. If TRUE, the log expression will be applied.}

\item{font.size.in.bubble}{The font size inside the bubble. Default 2}
}
\description{
Plot bubble-plot for gene expression per cluster
}
