% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_bubble_for_genes_per_custom_group_of_cells}
\alias{plot_bubble_for_genes_per_custom_group_of_cells}
\title{Plot bubble-plot for gene expression per custom group of cells}
\usage{
plot_bubble_for_genes_per_custom_group_of_cells(
  object,
  custom_group_of_cells = list(),
  gene_list = c(),
  point.color1 = "orange",
  point.color2 = "firebrick",
  buble.scale = 6,
  gene.font.size = 12,
  axist.x.font.size = 12,
  font.size.in.bubble = 2,
  show.percent = FALSE,
  log = T,
  IsApplyClustering = TRUE,
  IsClusterByRow = TRUE,
  clustering_method = "complete",
  IsClusterByColumn = TRUE,
  IsCustomOrder = FALSE,
  IsCustomOrderByRow = FALSE,
  IsCustomOrderByColumn = FALSE,
  row.custom.order = "",
  column.custom.order = ""
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{custom_group_of_cells}{The list of custome group of cells.}

\item{gene_list}{The vector of gene names.}

\item{point.color1}{The color for low expression level. Default orange}

\item{point.color2}{The color for high expression level. Default firebrick}

\item{buble.scale}{The scale size of the bubble. Default 6}

\item{gene.font.size}{The font size of gene name. Default 12}

\item{axist.x.font.size}{The label on the x-axis font size. Default 12}

\item{font.size.in.bubble}{The font size inside the bubble. Default 2}

\item{show.percent}{is logical. If TRUE, the percentage will be displayed.}

\item{log}{is logical. If TRUE, the log expression will be applied.}

\item{IsApplyClustering}{is logical. If TRUE, the bubble plot will be clustered by using HCL.}

\item{IsClusterByRow}{is logical. If TRUE, clustering by row.}

\item{clustering_method}{The clustering method. Default 'complete'}

\item{IsClusterByColumn}{is logical. If TRUE, clustering by column.}

\item{IsCustomOrder}{is logical, If TRUE, the input custom order of row and column will be used.}

\item{IsCustomOrderByRow}{is logical, If TRUE, row will be ordered by the custom input.}

\item{IsCustomOrderByColumn}{is logical, If TRUE, column will be ordered by the custom input.}

\item{row.custom.order}{is the vector of custom row names.}

\item{column.custom.order}{is the vector of custom column names.}
}
\description{
Plot bubble-plot for gene expression per custom group of cells
}
