% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_cells_annotation}
\alias{plot_cells_annotation}
\title{Plot cell annotation}
\usage{
plot_cells_annotation(object, type = c("histogram", "boxplot"))
}
\arguments{
\item{object}{The SingCellaR object.}

\item{type}{Type of plots including the histogram and boxplot.}
}
\description{
Plot cell annotation
}
