% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_diffusionmap_label_by_genes}
\alias{plot_diffusionmap_label_by_genes}
\title{Plot diffusion map with gene expression}
\usage{
plot_diffusionmap_label_by_genes(
  object,
  gene_list = c(),
  point.size = 0.3,
  point.color1 = "blue",
  point.color2 = "red",
  x_eigenVal = "DC1",
  y_eigenVal = "DC2"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gene_list}{The vector of gene names}

\item{point.size}{The point size. Default 0.3}

\item{point.color1}{The first color of gene expression gradient. Default blue}

\item{point.color2}{The second color of gene expression gradient. Default red}

\item{x_eigenVal}{The diffusion map dimension on x axis. Default DC1}

\item{y_eigenVal}{The diffusion map dimension on y axis. Default DC2}
}
\description{
Plot diffusion map with gene expression
}
