% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_diffusionmap_label_by_multiple_gene_sets}
\alias{plot_diffusionmap_label_by_multiple_gene_sets}
\title{Plot diffusion map with the expression of multiple gene sets}
\usage{
plot_diffusionmap_label_by_multiple_gene_sets(
  object,
  gmt.file = c(),
  show_gene_sets = c(),
  custom_color = c(),
  isNormalizedByHouseKeeping = T,
  x_eigenVal = "DC1",
  y_eigenVal = "DC2",
  point.size = 1,
  showLegend = T,
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gmt.file}{The GMT file for gene sets.}

\item{show_gene_sets}{The names of selected gene sets.}

\item{custom_color}{The custom color names.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{x_eigenVal}{The diffusion map dimension on x axis. Default DC1}

\item{y_eigenVal}{The diffusion map dimension on y axis. Default DC2}

\item{point.size}{The point size. Default 1}

\item{showLegend}{is logical. If TRUE, the figure legend will be shown.}

\item{background.color}{The background color of the plot. Default white}
}
\description{
Plot diffusion map with the expression of multiple gene sets
}
