% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_a_feature_of_interest}
\alias{plot_forceDirectedGraph_label_by_a_feature_of_interest}
\title{Plot force-directed graph with a feature of interest}
\usage{
plot_forceDirectedGraph_label_by_a_feature_of_interest(
  object,
  feature = "",
  title = "",
  vertex.size = 0.5,
  edge.size = 0.2,
  edge.color = "gray85",
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{feature}{The selected feature of interest.}

\item{title}{The title of the plot.}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default gray85}

\item{background.color}{The background color. Default white}

\item{vertext.size}{The node size. Default 0.5}
}
\description{
Plot force-directed graph with a feature of interest
}
