% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_a_signature_gene_set}
\alias{plot_forceDirectedGraph_label_by_a_signature_gene_set}
\title{Plot force-directed graph with the expression of a gene set}
\usage{
plot_forceDirectedGraph_label_by_a_signature_gene_set(
  object,
  gene_list = c(),
  gene_set_name = c(),
  isNormalizedByHouseKeeping = F,
  vertex.size = 0.5,
  vertex.color1 = "blue",
  vertex.color2 = "red",
  edge.size = 0.2,
  edge.color = "gray",
  vertex.base.color = "gray35",
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gene_list}{The vector of genes.}

\item{gene_set_name}{The name of a gene set.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{vertex.color1}{The first color of gene expression gradient. Default blue}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default gray}

\item{vertex.base.color}{The node base color. Default gray35}

\item{background.color}{The background color. Default white}

\item{vertext.size}{The node size. Default 0.5}

\item{vertext.color2}{The second color of gene expression gradient. Default red}
}
\description{
Plot force-directed graph with the expression of a gene set
}
