% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_AUCell_score}
\alias{plot_forceDirectedGraph_label_by_AUCell_score}
\title{Plot force-direct graph with AUCell score}
\usage{
plot_forceDirectedGraph_label_by_AUCell_score(
  object,
  AUCell_gene_set_name = c(),
  AUCell_score,
  AUCell_cutoff = 0,
  IsShowOnlySampleIDs = FALSE,
  selected.sampleID = c(),
  IsLimitedAUCscoreByClusters = FALSE,
  selected.limited.clusters = c(),
  clustering_method = "louvain",
  vertex.size = 1.5,
  edge.size = 0.2,
  vertex.color1 = "black",
  vertex.color2 = "red",
  edge.color = "gray",
  background.color = "white",
  showEdge = T
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_gene_set_name}{The vector of gene set names.}

\item{AUCell_score}{The dataframe of AUCell score per gene set.}

\item{AUCell_cutoff}{The cutoff score. Default 0}

\item{IsShowOnlySampleIDs}{is logical. If TRUE, only selected sample IDs will be shown.}

\item{selected.sampleID}{The selected sample IDs.}

\item{IsLimitedAUCscoreByClusters}{is logical. If TRUE, AUCell score will be limited by selected clusters.}

\item{selected.limited.clusters}{The selected clusters of interest.}

\item{clustering_method}{The clustering method.}

\item{vertex.size}{The node size. Default 1.5}

\item{edge.size}{The edge size. Default 0.2}

\item{vertex.color1}{The color for a low AUCell score. Default black}

\item{vertex.color2}{The color for a high AUCell score. Default red}

\item{edge.color}{The edge color.}

\item{background.color}{The background color.}

\item{showEdge}{is logical. If TRUE, the edge will be displayed.}
}
\description{
Plot force-direct graph with AUCell score
}
