% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_clusters}
\alias{plot_forceDirectedGraph_label_by_clusters}
\title{Plot force-directed graph with identified clusters}
\usage{
plot_forceDirectedGraph_label_by_clusters(
  object,
  show_method = c("walktrap", "louvain", "kmeans", "merged_walktrap", "merged_louvain",
    "merged_kmeans"),
  vertex.size = 2,
  mark.clusters = TRUE,
  mark.font.size = 10,
  mark.font.color = "yellow",
  add.cluster.in.cells = FALSE,
  cluster.font.size = 1,
  cluster.font.color = "yellow",
  edge.size = 0.2,
  edge.color = "gray",
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{show_method}{The clustering method names used to generate clusters.}

\item{mark.clusters}{is logical. If TRUE, the number of identified clusters will be displayed.}

\item{mark.font.size}{The font size of the cluster name. Default 10}

\item{mark.font.color}{The font color of the cluster name. Default yellow}

\item{add.cluster.in.cells}{is logical. If TRUE, the cluster name will be displayed in each data point.}

\item{cluster.font.size}{The font size of cluster name in each data point. Default 1}

\item{cluster.font.color}{The font color of cluster name in each data point. Default yellow}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default gray}

\item{background.color}{The background color. Default white}

\item{vertext.size}{The node size. Default 2}
}
\description{
Plot force-directed graph with identified clusters
}
