% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_multiple_gene_sets}
\alias{plot_forceDirectedGraph_label_by_multiple_gene_sets}
\title{Plot force-directed graph with the expression of selected multiple gene sets}
\usage{
plot_forceDirectedGraph_label_by_multiple_gene_sets(
  object,
  gmt.file = c(),
  show_gene_sets = c(),
  custom_color = c(),
  isNormalizedByHouseKeeping = T,
  vertex.size = 1.5,
  showEdge = T,
  edge.size = 0.2,
  edge.color = "gray",
  showLegend = T,
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gmt.file}{The GMT file for gene sets.}

\item{show_gene_sets}{The names of selected gene sets.}

\item{custom_color}{The custom color names.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{showEdge}{is logical. If TRUE, the graph's edges will be displayed.}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default gray}

\item{showLegend}{is logical. If TRUE, the figure legend will be displayed.}

\item{background.color}{The background color. Default white}

\item{vertext.size}{The node size. Default 1.5}
}
\description{
Plot force-directed graph with the expression of selected multiple gene sets
}
