% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_sampleID}
\alias{plot_forceDirectedGraph_label_by_sampleID}
\title{Plot force-directed graph with the sampleID}
\usage{
plot_forceDirectedGraph_label_by_sampleID(
  object,
  vertex.size = 0.5,
  edge.size = 0.2,
  edge.color = "gray",
  vertex.base.color = "gray35"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default gray45}

\item{vertex.base.color}{The node base color. Default gray35}

\item{vertext.size}{The node size. Default 0.5}
}
\description{
Plot force-directed graph with the sampleID
}
