% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_forceDirectedGraph_label_by_selected_sampleID}
\alias{plot_forceDirectedGraph_label_by_selected_sampleID}
\title{Plot force-directed graph with selected sampleIDs}
\usage{
plot_forceDirectedGraph_label_by_selected_sampleID(
  object,
  selected.sampleID = c(),
  title = "",
  vertex.size = 0.5,
  vertex.front.color = "red",
  edge.size = 0.2,
  edge.color = "white",
  vertex.base.color = "gray",
  alpha.base = 0.2
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{selected.sampleID}{The vector of selected sampleIDs.}

\item{title}{The title of the plot.}

\item{vertex.front.color}{The color of selected samples. Default red}

\item{edge.size}{The size of the edge. Default 0.2}

\item{edge.color}{The edge color. Default white}

\item{vertex.base.color}{The node base color. Default gray}

\item{alpha.base}{The alpha parameter of node. Default 0.2}

\item{vertext.size}{The node size. Default 0.5}
}
\description{
Plot force-directed graph with selected sampleIDs
}
