% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_heatmap_for_differential_genes}
\alias{plot_heatmap_for_differential_genes}
\title{Plot heatmap for differential genes}
\usage{
plot_heatmap_for_differential_genes(
  objectA = objectA,
  objectB = objectB,
  cellsA = c(),
  cellsB = c(),
  gene_list = c(),
  groupA.name = "",
  groupB.name = "",
  show.cell_annotation = c(),
  isClusterByRow = T,
  isClusterByCol = T,
  distance_row = "euclidean",
  distance_column = "euclidean",
  show_row_dend = T,
  show_column_dend = T,
  col.scaled.min = -2.5,
  col.scaled.max = 2.5,
  col.low = "blue",
  col.mid = "black",
  col.high = "red",
  rowFont.size = 6,
  use_raster = FALSE
)
}
\arguments{
\item{objectA}{The SingCellaR objectA.}

\item{objectB}{The SingCellaR objectB.}

\item{cellsA}{The vector of cell names in group A.}

\item{cellsB}{The vector of cell names in group B.}

\item{gene_list}{The vector of gene names.}

\item{groupA.name}{The group A cell names.}

\item{groupB.name}{The group B cell names.}

\item{show.cell_annotation}{The vector of cell annotation names}

\item{isClusterByRow}{is logical. If TRUE, the clustering by row will be performed.}

\item{isClusterByCol}{is logical. If TRUE, the clustering by column will be performed.}

\item{distance_row}{The distance method for row. Default euclidean}

\item{distance_column}{The distance method for column. Default euclidean}

\item{show_row_dend}{is logical. If TRUE, the row dendrogram will be shown.}

\item{show_column_dend}{is logical. If TRUE, the column dendrogram will be shown.}

\item{col.scaled.min}{The minimum scaled value. Default -2.5}

\item{col.scaled.max}{The maximum scaled value. Default 2.5}

\item{col.low}{The low color gradient. Default blue}

\item{col.mid}{The mid color gradient. Default black}

\item{col.high}{The high color gradient. Default red}

\item{rowFont.size}{The row font size. Default 6}

\item{use_raster}{Whether render the heatmap body as a raster image. It helps to reduce file size when the matrix is huge.}
}
\description{
Plot heatmap for differential genes
}
