% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_heatmap_for_fGSEA_all_clusters}
\alias{plot_heatmap_for_fGSEA_all_clusters}
\title{Plot heatmap for GSEA for all comparisons}
\usage{
plot_heatmap_for_fGSEA_all_clusters(
  fGSEA_results.data.frame,
  isApplyCutoff = F,
  adjusted_pval = 0.25,
  add_small_value = 1e-04,
  use_pvalues_for_clustering = T,
  show_NES_score = T,
  show_only_NES_positive_score = F,
  show_only_NES_negative_score = F,
  custom_order_samples = c(),
  fontsize_row = 5,
  format.digits = 2,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  clustering_method = "complete",
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  gaps_row = NULL,
  gaps_col = NULL,
  show_text_for_ns = T
)
}
\arguments{
\item{fGSEA_results.data.frame}{The input dataframe for fGSEA analysis results.}

\item{isApplyCutoff}{is logical. If TRUE, the cutoff will be applied.}

\item{adjusted_pval}{The cutoff adjusted p-value. Default 0.25}

\item{use_pvalues_for_clustering}{is logical. If TRUE, the p-values will be used for clustering analysis.}

\item{show_NES_score}{is logical. If TRUE, the enrichment score will be displayed.}

\item{show_only_NES_positive_score}{is logical. If TRUE, the only NES positive scores will be displayed.}

\item{show_only_NES_negative_score}{is logical. If TRUE, the only NES negative scores will be displayed.}

\item{custom_order_samples}{The vector of input custom order of GSEA comparison on the heatmap.}

\item{fontsize_row}{The row font size. Default 5}

\item{format.digits}{The format of digit number shows on the heatmap. Default 2}

\item{cluster_rows}{is logical. If TRUE, the clustering analysis will be performed by row.}

\item{cluster_cols}{is logical. If TRUE, the clustering analysis will be performed by column.}

\item{clustering_method}{The HCL clustering method. Default complete}

\item{clustering_distance_rows}{The distance of rows for the clustering. Default euclidean}

\item{clustering_distance_cols}{The distance of columns for the clustering. Default euclidean.}

\item{gaps_row}{The vector of gaps for row.}

\item{gaps_col}{The vector of gaps for column.}

\item{show_text_for_ns}{is logical. If TRUE. The ns or 'non-significant' will be displayed.}

\item{object}{The SingCellaR object.}
}
\description{
Plot heatmap for GSEA for all comparisons
}
