% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_heatmap_for_marker_genes}
\alias{plot_heatmap_for_marker_genes}
\title{Plot heatmap for identified marker genes}
\usage{
plot_heatmap_for_marker_genes(
  object,
  cluster.type = c("walktrap", "louvain", "kmeans", "merged_louvain",
    "merged_walktrap", "merged_kmeans"),
  n.TopGenes = 5,
  min.log2FC = 0.5,
  min.expFraction = 0.3,
  isClusterByRow = F,
  col.scaled.min = -2.5,
  col.scaled.max = 2.5,
  col.low = "blue",
  col.mid = "black",
  col.high = "red",
  rowFont.size = 6,
  split.line.col = "white",
  split.line.type = 1,
  split.line.lwd = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{n.TopGenes}{The number of top differential genes. Default 5}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.5}

\item{min.expFraction}{The minimum cutoff for the fraction of expressing cell frequency. Default 0.3}

\item{isClusterByRow}{is logical. If TRUE, the clustering by row will be performed.}

\item{col.scaled.min}{The minimum scaled value. Default -2.5}

\item{col.scaled.max}{The maximum scaled value. Default 2.5}

\item{col.low}{The low color gradient. Default blue}

\item{col.mid}{The mid color gradient. Default black}

\item{col.high}{The high color gradient. Default red}

\item{rowFont.size}{The row font size. Default 6}

\item{split.line.col}{The split line color. Default white}

\item{split.line.type}{The split line type. Default 1}

\item{split.line.lwd}{The split line lwd. Default 1}
}
\description{
Plot heatmap for identified marker genes
}
