% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_jaccard_similarity_among_clusters}
\alias{plot_jaccard_similarity_among_clusters}
\title{Plot jaccard similar across indetified clusters}
\usage{
plot_jaccard_similarity_among_clusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  min.log2FC = 0.3,
  min.expFraction = 0.3
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{min.log2FC}{The minimum log2FC cutoff. Default 0.3}

\item{min.expFraction}{The minimum fraction of expressing cell quency. Default 0.3}
}
\description{
Plot jaccard similar across indetified clusters
}
