% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_PCA_Elbowplot}
\alias{plot_PCA_Elbowplot}
\title{Plot PCA's elbow plot}
\usage{
plot_PCA_Elbowplot(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Plot PCA's elbow plot
}
