% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_ShortestPath_on_3D_knn_graph}
\alias{plot_ShortestPath_on_3D_knn_graph}
\title{Plot the shortest-path on the 3D KNN-Graph}
\usage{
plot_ShortestPath_on_3D_knn_graph(
  object,
  start_cluster = "",
  end_cluster = "",
  cluster.type = c("louvain", "walktrap", "kmeans"),
  path.color = "red",
  vertext.size = 0.4,
  edge.color = "gray45"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{start_cluster}{The starting cluster.}

\item{end_cluster}{The ending cluster.}

\item{cluster.type}{The clustering method name.}

\item{path.color}{The path color. Default red}

\item{vertext.size}{The node size. Default 0.4}

\item{edge.color}{The edge color. Default gray}
}
\description{
Plot the shortest-path on the 3D KNN-Graph
}
