% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_tsne_label_by_a_feature_of_interest}
\alias{plot_tsne_label_by_a_feature_of_interest}
\title{Plot TSNE and display only the feature of interest}
\usage{
plot_tsne_label_by_a_feature_of_interest(object, feature = "", point.size = 1)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{feature}{The feature name. The features can be found in the cell metadata.}

\item{point.size}{The point size. Default 1}
}
\description{
Plot TSNE and display only the feature of interest
}
