% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_tsne_label_by_multiple_gene_sets_with_limited_sampleID}
\alias{plot_tsne_label_by_multiple_gene_sets_with_limited_sampleID}
\title{Plot TSNE with the expression of multiple gene sets on selected sample IDs}
\usage{
plot_tsne_label_by_multiple_gene_sets_with_limited_sampleID(
  object,
  gmt.file = c(),
  selected.sampleID = c(),
  title = "",
  show_gene_sets = c(),
  custom_color = c(),
  isNormalizedByHouseKeeping = T,
  IsDownsample = F,
  downsample.size = 0,
  title.color = "red",
  title.font.size = 16,
  point.size = 2,
  showLegend = T,
  background.color = "white"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gmt.file}{The GMT file.}

\item{selected.sampleID}{Selected sample IDs}

\item{show_gene_sets}{The name of gene sets to be shown.}

\item{custom_color}{The vector of cutom color to specify gene sets.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{IsDownsample}{is logical. If TRUE, the cells will be downsampled.}

\item{downsample.size}{The number of cells for the downsample. Default 0}

\item{title.color}{The color of the title. Default red}

\item{title.font.size}{The title font size. Default 16}

\item{point.size}{The point size. Default 2}

\item{showLegend}{is logical. If TRUE, the figure legend will be displayed.}

\item{background.color}{The background color. Default white}
}
\description{
Plot TSNE with the expression of multiple gene sets on selected sample IDs
}
