% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_tsne_label_by_qc}
\alias{plot_tsne_label_by_qc}
\title{Plot TSNE with QC information}
\usage{
plot_tsne_label_by_qc(object, point.size = 0.2)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{point.size}{The point size}
}
\description{
Plot TSNE with QC information
}
