% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_tsne_label_by_sampleID}
\alias{plot_tsne_label_by_sampleID}
\title{Plot TSNE with sampleID information}
\usage{
plot_tsne_label_by_sampleID(object, point.size = 0.2)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{point.size}{The point size. Default 0.2}
}
\description{
Plot TSNE with sampleID information
}
