% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_a_feature_of_interest}
\alias{plot_umap_label_by_a_feature_of_interest}
\title{Plot UMAP and display only the feature of interest}
\usage{
plot_umap_label_by_a_feature_of_interest(
  object,
  feature = "",
  point.size = 1,
  mark.feature = TRUE,
  mark.font.size = 5,
  mark.font.color = "black"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{feature}{The feature name. The features can be found in the cell metadata.}

\item{point.size}{The point size. Default 1}

\item{mark.feature}{showing a selected feature name. Default TRUE}

\item{mark.font.size}{the size of the feature name. Default 5}

\item{mark.font.color}{the color of the feature name. Default black}
}
\description{
Plot UMAP and display only the feature of interest
}
