% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_a_signature_gene_set}
\alias{plot_umap_label_by_a_signature_gene_set}
\title{Plot UMAP with the expression of selected gene set}
\usage{
plot_umap_label_by_a_signature_gene_set(
  object,
  gene_list = c(),
  gene_set_name = c(),
  isNormalizedByHouseKeeping = F,
  point.size = 1,
  point.color1 = "black",
  point.color2 = "red"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gene_list}{The vector of gene names.}

\item{gene_set_name}{The name of gene set.}

\item{isNormalizedByHouseKeeping}{is logical. If TRUE, the gene expression will be normalized by the expression of house keeping genes.}

\item{point.size}{The point size. Default 0.3}

\item{point.color1}{The first color of gene expression gradient. Default blue}

\item{point.color2}{The second color of gene expression gradient. Default red}
}
\description{
Plot UMAP with the expression of selected gene set
}
