% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_AUCell_score}
\alias{plot_umap_label_by_AUCell_score}
\title{Plot UMAP with AUCell score}
\usage{
plot_umap_label_by_AUCell_score(
  object,
  AUCell_gene_set_name = c(),
  AUCell_score,
  AUCell_cutoff = 0,
  IsShowOnlySampleIDs = FALSE,
  selected.sampleID = c(),
  IsLimitedAUCscoreByClusters = FALSE,
  selected.limited.clusters = c(),
  clustering_method = "louvain",
  point.size = 1,
  point.color1 = "black",
  point.color2 = "red"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_gene_set_name}{The vector of gene set names.}

\item{AUCell_score}{The dataframe of AUCell score per gene set.}

\item{AUCell_cutoff}{The cutoff score. Default 0}

\item{IsShowOnlySampleIDs}{is logical. If TRUE, only selected sample IDs will be shown.}

\item{selected.sampleID}{The selected sample IDs.}

\item{IsLimitedAUCscoreByClusters}{is logical. If TRUE, AUCell score will be limited by selected clusters.}

\item{selected.limited.clusters}{The selected clusters of interest.}

\item{clustering_method}{The clustering method.}

\item{point.size}{The point size. Default 1}

\item{point.color1}{The color for a low AUCell score. Default gray}

\item{point.color2}{The color for a high AUCell score . Default red}
}
\description{
Plot UMAP with AUCell score
}
