% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_genes}
\alias{plot_umap_label_by_genes}
\title{Plot UMAP with gene expression}
\usage{
plot_umap_label_by_genes(
  object,
  gene_list = c(),
  point.size = 0.3,
  point.color1 = "blue",
  point.color2 = "red"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gene_list}{The vector of gene names}

\item{point.size}{The point size. Default 0.3}

\item{point.color1}{The first color of gene expression gradient. Default blue}

\item{point.color2}{The second color of gene expression gradient. Default red}
}
\description{
Plot UMAP with gene expression
}
