% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_qc}
\alias{plot_umap_label_by_qc}
\title{Plot UMAP with QC information}
\usage{
plot_umap_label_by_qc(object, point.size = 0.2)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{point.size}{The point size. Default 0.2}
}
\description{
Plot UMAP with QC information
}
