% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_umap_label_by_selected_sampleID}
\alias{plot_umap_label_by_selected_sampleID}
\title{Plot UMAP and display only selected sampleIDs}
\usage{
plot_umap_label_by_selected_sampleID(
  object,
  selected.sampleID = c(),
  title = "",
  point.size = 0.5,
  point.front.color = "red",
  point.base.color = "gray",
  alpha.base = 0.2,
  title.color = "red",
  title.font.size = 16
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{selected.sampleID}{The vector of selected sampleIDs.}

\item{title}{The title of the plot.}

\item{point.size}{The point size. Default 0.5}

\item{point.front.color}{The point color of selected sampleIDs. Default red}

\item{point.base.color}{The point color of unselected samplesIDs.}

\item{alpha.base}{The alpha parameter. Default 0.2}

\item{title.color}{The title color. Default red}

\item{title.font.size}{The font size of the title. Default 16}
}
\description{
Plot UMAP and display only selected sampleIDs
}
