% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_UMIs_of_HouseKeepingGenes}
\alias{plot_UMIs_of_HouseKeepingGenes}
\title{Plot UMIs of house keeping genes}
\usage{
plot_UMIs_of_HouseKeepingGenes(
  object,
  selected_top_genes = 100,
  selected_top_cells = 300
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{selected_top_genes}{The number of top genes. Default 100}

\item{selected_top_cells}{The number of top cells. Default 300}
}
\description{
Plot UMIs of house keeping genes
}
