% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_UMIs_vs_Detected_genes}
\alias{plot_UMIs_vs_Detected_genes}
\title{Plot UMI count vs detected genes}
\usage{
plot_UMIs_vs_Detected_genes(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Plot UMI count vs detected genes
}
