% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_variable_genes}
\alias{plot_variable_genes}
\title{Plot highly variable genes}
\usage{
plot_variable_genes(
  object,
  quantile_genes_expr_for_fitting = 0.2,
  quantile_genes_cv2_for_fitting = 0.8
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{quantile_genes_expr_for_fitting}{The minimum quantile of gene expression used for model fitting. Default 0.2}

\item{quantile_genes_cv2_for_fitting}{The maximum quantile of coefficient of variation used for model fitting. Default 0.8}
}
\description{
Plot highly variable genes
}
