% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_violin_for_differential_genes}
\alias{plot_violin_for_differential_genes}
\title{Plot violin-plot for differential genes}
\usage{
plot_violin_for_differential_genes(
  objectA = objectA,
  objectB = objectB,
  cellsA = c(),
  cellsB = c(),
  gene_list = c(),
  take_log2 = T,
  groupA.name = "",
  groupB.name = "",
  xlab.text.size = 5,
  point.size = 0.5,
  point.alpha = 0.1
)
}
\arguments{
\item{objectA}{The SingCellaR objectA.}

\item{objectB}{The SingCellaR objectB.}

\item{cellsA}{The vector of cell names in group A.}

\item{cellsB}{The vector of cell names in group B.}

\item{gene_list}{The vector of gene names.}

\item{take_log2}{is logical. If TRUE, the expression shows in the log2 scale.}

\item{groupA.name}{The group A name.}

\item{groupB.name}{The group B name.}

\item{xlab.text.size}{The font size of the label in the x-axis. Default 5}

\item{point.size}{The point size. Default 0.2}

\item{point.alpha}{The alpha parameter. Default 0.1}
}
\description{
Plot violin-plot for differential genes
}
