% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellPlots.R
\name{plot_violin_for_marker_genes}
\alias{plot_violin_for_marker_genes}
\title{Plot violin-plot for marker genes}
\usage{
plot_violin_for_marker_genes(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  gene_list = c(),
  take_log2 = T,
  xlab.text.size = 5,
  point.size = 0.2,
  point.alpha = 0.1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{gene_list}{The vector of gene names.}

\item{take_log2}{is logical. If TRUE, the expression shows in the log2 scale.}

\item{xlab.text.size}{The font size of the label in the x-axis. Default 5}

\item{point.size}{The point size. Default 0.2}

\item{point.alpha}{The alpha parameter. Default 0.1}
}
\description{
Plot violin-plot for marker genes
}
