% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{preprocess_integration_for_ADT}
\alias{preprocess_integration_for_ADT}
\title{Preprocessing for ADT}
\usage{
preprocess_integration_for_ADT(object)
}
\arguments{
\item{object}{The SingCellaR object.}
}
\description{
Preprocessing for ADT
}
