% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{process_cells_annotation}
\alias{process_cells_annotation}
\title{Process all cells annotation}
\usage{
process_cells_annotation(object, mitochondiral_genes_start_with = "MT-")
}
\arguments{
\item{object}{The SingCellaR object.}

\item{mitochondiral_genes_start_with}{The unique prefix alphabets for mitocondrial genes such as 'MT-' for human and 'mt-' for mouse genes.}
}
\description{
Process all cells annotation
}
