% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{remove_clusters}
\alias{remove_clusters}
\title{Remove clusters}
\usage{
remove_clusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans"),
  cluster_ids = c()
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{cluster_ids}{The vector of cluster ids that will be removed e.g. c('cl1:cl2','cl9:cl10').}
}
\description{
Remove clusters
}
