% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{remove_unwanted_confounders}
\alias{remove_unwanted_confounders}
\title{Removing unwanted confounders using limma}
\usage{
remove_unwanted_confounders(
  object,
  residualModelFormulaStr = "~UMI_count+percent_mito",
  preserved_feature = "",
  block.gene.size = 2000
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{residualModelFormulaStr}{The model for removing confounder variables.}

\item{preserved_feature}{is a defined preserved variable such as a cell genotype.}

\item{block.gene.size}{is the number of genes in each processing block.}
}
\description{
Removing unwanted confounders using limma
}
