% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{remove_unwanted_genes_from_variable_gene_set}
\alias{remove_unwanted_genes_from_variable_gene_set}
\title{Remove unwanted genes from identified list of variable genes}
\usage{
remove_unwanted_genes_from_variable_gene_set(
  object,
  gmt.file = c(),
  removed_gene_sets = c("Ribosomal_gene", "Mitochondrial_gene")
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{gmt.file}{The GMT file that containes the signature gene sets.}

\item{removed_gene_sets}{The vector of gene sets to be removed from the list of variable genes.}
}
\description{
Remove unwanted genes from identified list of variable genes
}
