% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Run_AUCell}
\alias{Run_AUCell}
\title{Run AUCell}
\usage{
Run_AUCell(
  object,
  AUCell_buildRankings.file = "",
  geneSets.gmt.file = "",
  n.random.genes = 500
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_buildRankings.file}{The input file name of AUCell rankings.}

\item{geneSets.gmt.file}{The GMT file name that contains gene sets.}

\item{n.random.genes}{The number of a random gene set.}
}
\description{
Run AUCell
}
