% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Run_fGSEA_analysis}
\alias{Run_fGSEA_analysis}
\title{Run GSEA analysis}
\usage{
Run_fGSEA_analysis(
  objectA = objectA,
  objectB = objectB,
  cellsA = c(),
  cellsB = c(),
  GSEAPrerankedGenes_info = "",
  gmt.file = "",
  plotGSEA = T,
  nTopGenesets = 10,
  minSize = 5,
  maxSize = 1000,
  eps = 0,
  nPermSimple = 1000
)
}
\arguments{
\item{objectA}{The SingCellaR object A}

\item{objectB}{The SingCellaR object B}

\item{cellsA}{The vector of cell names from object A}

\item{cellsB}{The vector of cell names from object B}

\item{GSEAPrerankedGenes_info}{The dataframe object of preranked genes.}

\item{gmt.file}{The GMT file name}

\item{plotGSEA}{is logical. If TRUE, this function will plot GSEA enrichment.}

\item{nTopGenesets}{The number of showing top gene sets.}

\item{minSize}{The cutoff minimum number of genes in each gene set. Gene set that contains the number of genes lower than this number will be excluded. Default 5}

\item{maxSize}{The cutoff maxiumum number of genes in each gene set. Gene set that contains the number of genes higher than this number will be excluded. Default 1000}

\item{eps}{The eps paramenter for fgsea, this parameter sets the boundary for calculating the p value. Default 0}

\item{nPermSimple}{The number of permutations in the simple fgsea implementation for preliminary estimation of P-values.}
}
\description{
Run GSEA analysis
}
