% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Run_fGSEA_for_a_selected_cluster_vs_the_rest_of_clusters}
\alias{Run_fGSEA_for_a_selected_cluster_vs_the_rest_of_clusters}
\title{Run GSEA analysis for a selected cluster vs the rest of clusters}
\usage{
Run_fGSEA_for_a_selected_cluster_vs_the_rest_of_clusters(
  object,
  cluster.id = "",
  cluster.type = c("louvain", "walktrap", "kmeans", "merged_walktrap",
    "merged_louvain", "merged_kmeans"),
  diff.gene.method = "wilcoxon",
  gmt.file = "",
  fishers_exact_test = 0.1,
  plotGSEA = T,
  nTopGenesets = 10,
  minSize = 5,
  maxSize = 2500,
  eps = 1e-10
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.id}{The specified cluster id (e.g. 'cluster1').}

\item{cluster.type}{The type of clustering method.}

\item{diff.gene.method}{The method for differential gene expression analysis. Default 'wilcoxon'.}

\item{gmt.file}{The GMT file name.}

\item{fishers_exact_test}{The fisher's exact test cutoff p-value.}

\item{plotGSEA}{is logical. If TRUE, this function will plot GSEA enrichment.}

\item{nTopGenesets}{The number of showing top gene sets.}

\item{minSize}{The cutoff minimum number of genes in each gene set. Gene set that contains the number of genes lower than this number will be excluded. Default 5}

\item{maxSize}{The cutoff maxiumum number of genes in each gene set. Gene set that contains the number of genes higher than this number will be excluded. Default 2500}

\item{eps}{The eps paramenter for fgsea, this parameter sets the boundary for calculating the p value. Default 1e-10}
}
\description{
Run GSEA analysis for a selected cluster vs the rest of clusters
}
