% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runCombat}
\alias{runCombat}
\title{Run combat integration}
\usage{
runCombat(
  object,
  use.reduced_dim = T,
  dim_reduction_method = c("pca", "nnmf"),
  n.dims.use = 30,
  batch_identifier = ""
)
}
\arguments{
\item{object}{The SingCellaR_Int object.}

\item{use.reduced_dim}{is logical. If TRUE, the reduced dimensions from PCA or nnmf will be used as the input. If FALSE, the normalized UMI will be used.}

\item{dim_reduction_method}{The method name for dimensional reduction.}

\item{n.dims.use}{The number of dimensions to be used. Default 30}

\item{batch_identifier}{The indicated feature name to separate the batch.}
}
\description{
Run combat integration
}
