% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\title{Diffusion map analysis}
\usage{
runDiffusionMap(
  object,
  dim_reduction_method = c("pca", "nnmf"),
  useIntegrativeEmbeddings = F,
  integrative_method = c("combat", "seurat", "harmony", "supervised_harmony", "liger"),
  n.dims.use = 50,
  n.dims.embed = 3,
  n.neighbors = 30,
  distance = c("euclidean", "cosine"),
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{dim_reduction_method}{Dimensional reduction method name that specifies the embedding matrix.}

\item{useIntegrativeEmbeddings}{is logical, if TRUE the embedding matrix genereated from integrative analysis will be used.}

\item{integrative_method}{The name of an integrative method.}

\item{n.dims.use}{The number of dimensions as the input.}

\item{n.dims.embed}{is the number of output dimension. Default 3}

\item{n.neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Default 30}

\item{distance}{The distance metric.}

\item{n.seed}{The number of seed.}
}
\description{
Diffusion map analysis
}
