% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runFA2_ForceDirectedGraph}
\alias{runFA2_ForceDirectedGraph}
\title{ForceAtlas2  analysis}
\usage{
runFA2_ForceDirectedGraph(
  object,
  dim_reduction_method = c("pca", "nnmf", "lsi"),
  useIntegrativeEmbeddings = F,
  integrative_method = c("combat", "seurat", "harmony", "supervised_harmony", "liger"),
  knn.metric = c("euclidean", "cosine"),
  n.dims.use = 30,
  n.neighbors = 5,
  n.seed = 1,
  knn.n_trees = 50,
  knn.n_threads = 1,
  fa2_n_iter = 1000,
  fa2_edgeWeightInfluence = 1,
  fa2_barnesHutTheta = 2,
  fa2_scalingRatio = 1,
  fa2_gravity = 0.05,
  fa2_jitterTolerance = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{useIntegrativeEmbeddings}{is logical, if TRUE the embedding matrix genereated from integrative analysis will be used.}

\item{integrative_method}{The name of an integrative method.}

\item{knn.metric}{Type of distance metric to use to find nearest neighbors.}

\item{n.dims.use}{The number of dimensions as the input. Default 30}

\item{n.neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Default 30}

\item{n.seed}{The number of set seed.}

\item{knn.n_trees}{The number of trees for building the annoy index. Default 50}

\item{knn.n_threads}{The number of threads. Default 1}

\item{fa2_n_iter}{The FA2 number of iterations. Default 1000}

\item{fa2_edgeWeightInfluence}{The FA2 edgeWeightInfluence parameter. Default 1}

\item{fa2_barnesHutTheta}{The FA2 barnesHutTheta paramenter. Default 2}

\item{fa2_scalingRatio}{The FA2 scalingRatio parameter. Default 1}

\item{fa2_gravity}{The FA2 gravity parameter. Default 0.05}

\item{fa2_jitterTolerance}{The FA2 jitterTolerance parameter. Default 1}
}
\description{
ForceAtlas2  analysis
}
