% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runHarmony}
\alias{runHarmony}
\title{Run harmony integration}
\usage{
runHarmony(
  object,
  n.dims.use = 30,
  covariates = c("data_set"),
  harmony.sigma = 0.1,
  harmony.nclust = NULL,
  harmony.tau = 0,
  harmony.block.size = 0.05,
  harmony.max.iter = 10,
  harmony.max.iter.cluster = 200,
  harmony.epsilon.cluster = 1e-05,
  harmony.epsilon.harmony = 1e-04,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR_Int object.}

\item{n.dims.use}{The number of PCA dimensions used for the input for harmony.}

\item{covariates}{The unwanted source of variations (e.g. batch, sample_id, etc).}

\item{harmony.sigma}{The harmony sigma parameter. Default 0.1}

\item{harmony.nclust}{The harmony nclust parameter. Default NULL}

\item{harmony.tau}{The harmony tau parameter. Default 0}

\item{harmony.block.size}{The harmony block.size parameter. Default 0.05}

\item{harmony.max.iter}{The harmony max iteration parameter. Default 10}

\item{harmony.max.iter.cluster}{The harmony max iteration cluster. Default 200}

\item{harmony.epsilon.cluster}{The harmony epsilon.cluster parameter. Default 1e-05}

\item{harmony.epsilon.harmony}{The harmony epsilon parameter. Default 1e-04}

\item{n.seed}{The set seed number.}
}
\description{
Run harmony integration
}
