% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellUtils.R
\name{runKBET}
\alias{runKBET}
\title{kBET analysis}
\usage{
runKBET(
  covariate_variable_name = "donor",
  reference.celltypes.rds.file = "",
  integrative.umap.rds.files = "",
  integrative.method.names = "",
  n.sample = 1000
)
}
\arguments{
\item{covariate_variable_name}{The batch effect variable of interest such as 'batch' and 'donor'. Default 'donor'}

\item{reference.celltypes.rds.file}{The RDS file name that contains cell type information.}

\item{integrative.umap.rds.files}{The vector of file names that contain UMAP coordinate information generated by different integrative methods.}

\item{integrative.method.names}{The vector of integrative method names represented in the same order as in 'integrative.umap.rds.files'.}

\item{n.sample}{The downsample size of data points used in kBET analysis.}
}
\description{
kBET analysis
}
