% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runKNN_Graph}
\alias{runKNN_Graph}
\title{K-nearest neighbor graph analysis}
\usage{
runKNN_Graph(
  object,
  dim_reduction_method = c("pca", "nnmf"),
  useIntegrativeEmbeddings = F,
  integrative_method = c("combat", "seurat", "harmony", "supervised_harmony"),
  n.dims.use = 30,
  n.neighbors = 5,
  knn.metric = c("euclidean", "cosine"),
  knn.n_trees = 50,
  knn.n_threads = 1,
  niter = 500L,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{useIntegrativeEmbeddings}{is logical, if TRUE the embedding matrix genereated from integrative analysis will be used.}

\item{integrative_method}{The name of an integrative method.}

\item{n.dims.use}{The number of dimensions as the input. Default 30}

\item{n.neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Default 30}

\item{knn.metric}{Type of distance metric to use to find nearest neighbors.}

\item{knn.n_trees}{The number of trees for building the annoy index. Default 50}

\item{knn.n_threads}{The number of threads. Default 1}

\item{niter}{Integer scalar, the number of iterations to perform.}

\item{n.seed}{The number of set seed.}
}
\description{
K-nearest neighbor graph analysis
}
