% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runLiger}
\alias{runLiger}
\title{Run liger integration}
\usage{
runLiger(object, liger.k = 30, liger.resolution = 0.4)
}
\arguments{
\item{object}{The SingCellaR_Int object.}

\item{liger.k}{liger's k number parameter. Default 30}

\item{liger.resolution}{liger's resolution parameter. Default 0.4}
}
\description{
Run liger integration
}
