% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellUtils.R
\name{runLISI}
\alias{runLISI}
\title{LISI analysis}
\usage{
runLISI(
  lisi_label1 = "donor",
  lisi_label2 = "CellType",
  reference.celltypes.rds.file = "",
  integrative.umap.rds.files = "",
  integrative.method.names = "",
  point.size = 5,
  IsShowPlot = TRUE
)
}
\arguments{
\item{lisi_label1}{The batch effect variable of interest such as 'batch' and 'donor'. Default 'donor'}

\item{lisi_label2}{The variable name that represents ground truth or high AUC score cell type'. Default 'CellType'}

\item{reference.celltypes.rds.file}{The RDS file name that contains cell type information.}

\item{integrative.umap.rds.files}{The vector of file names that contain UMAP coordinate information generated by different integrative methods.}

\item{integrative.method.names}{The vector of integrative method names represented in the same order as in 'integrative.umap.rds.files'.}

\item{point.size}{The size of data points in the plot.}

\item{IsShowPlot}{If TRUE, the function will show the plot. If not, it will return the lisi's scores.}
}
\description{
LISI analysis
}
