% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runNNMF}
\alias{runNNMF}
\title{NNMF analysis}
\usage{
runNNMF(
  object,
  k = 30,
  use.regressout.data = T,
  use.scanorama.integrative.matrix = F,
  n.threads = 4,
  max.iter = 1000,
  rel.tol = 1e-05
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{use.regressout.data}{is logical, if TRUE, NNMF uses the regressout data. If FALSE, NNMF uses the normalized data without removing any umwanted source of variations.}

\item{use.scanorama.integrative.matrix}{is logical, if TRUE scannorama integrative matrix will be used.}

\item{n.threads}{The number of threads for running. Default 4}

\item{max.iter}{The number of NNMF interations. Default 1000}

\item{rel.tol}{The nnmf parameter. Default 1e-05}
}
\description{
NNMF analysis
}
