% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runPCA}
\alias{runPCA}
\title{PCA analysis}
\usage{
runPCA(
  object,
  use.components = 50,
  use.regressout.data = T,
  use.scanorama.integrative.matrix = F
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{use.components}{The number of principal components. Default 50}

\item{use.scanorama.integrative.matrix}{is the logical, if TRUE scannorama integrative matrix will be used.}

\item{regressout.data}{is logical, if TRUE, PCA uses the regressout data. If FALSE, PCA uses the normalized data without removing any umwanted source of variations.}
}
\description{
PCA analysis
}
