% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runScanorama}
\alias{runScanorama}
\title{Run scanorama integration}
\usage{
runScanorama(object, useCombinedVarGenesFromIndividualSample = F)
}
\arguments{
\item{object}{The SingCellaR_Int object.}

\item{useCombinedVarGenesFromIndividualSample}{is logical. If TRUE, all combined variable genes from individual sample will be used. If FALSE, highly variable genes from the whole combined samples will be used.}
}
\description{
Run scanorama integration
}
