% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runSeuratIntegration}
\alias{runSeuratIntegration}
\title{Run Seurat integration}
\usage{
runSeuratIntegration(
  object,
  Seurat.metadata = "",
  Seurat.split.by = "",
  n.dims.use = 30,
  Seurat.variablegenes.method = "vst",
  Seurat.variablegenes.number = 2000,
  use.SingCellaR.varGenes = F,
  k.anchor = 5,
  k.filter = 200,
  k.score = 30
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{Seurat.metadata}{cell metadata (a data frame format)}

\item{Seurat.split.by}{The indicated feature for splitting samples for the integration.}

\item{n.dims.use}{The number of PCA dimensions used for the input for Seurat.}

\item{Seurat.variablegenes.method}{The method for variable genes selection. Default 'vst'}

\item{Seurat.variablegenes.number}{The number of highly varible genes. Default 2000}

\item{use.SingCellaR.varGenes}{is logical. If TRUE, the highly variable genes identified by SingCellaR will be used.}

\item{k.anchor}{Seurat's k anchor parameter. Default 5}

\item{k.filter}{Seurat's k filter parameter. Default 200}

\item{k.score}{Seurat's k score parameter. Default 30}
}
\description{
Run Seurat integration
}
