% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runSeuratIntegration_with_rpca}
\alias{runSeuratIntegration_with_rpca}
\title{Run Seurat integration using reciprocal PCA (RPCA)}
\usage{
runSeuratIntegration_with_rpca(
  object,
  Seurat.metadata = "",
  Seurat.split.by = "",
  n.dims.use = 30,
  Seurat.variablegenes.method = "vst",
  Seurat.variablegenes.number = 2000,
  use.SingCellaR.varGenes = F
)
}
\arguments{
\item{object}{The SingCellaR_Int object.}

\item{Seurat.metadata}{cell metadata (a data frame format)}

\item{Seurat.split.by}{The indicated feature for splitting samples for the integration.}

\item{n.dims.use}{The number of PCA dimensions used for the input for Seurat.}

\item{Seurat.variablegenes.method}{The method for variable genes selection. Default 'vst'}

\item{Seurat.variablegenes.number}{The number of highly varible genes. Default 2000}

\item{use.SingCellaR.varGenes}{is logical. If TRUE, the highly variable genes identified by SingCellaR will be used.}
}
\description{
Run Seurat integration using reciprocal PCA (RPCA)
}
