% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runSupervised_Harmony}
\alias{runSupervised_Harmony}
\title{Run supervised harmony integration}
\usage{
runSupervised_Harmony(
  object,
  n.dims.use = 30,
  fGSEA.minSize = 10,
  fGSEA.maxSize = 500,
  fGSEA.eps = 1e-10,
  hcl.height.cutoff = 0.25,
  covariates = c("data_set"),
  harmony.sigma = 0.1,
  harmony.tau = 0,
  harmony.block.size = 0.05,
  harmony.max.iter = 10,
  harmony.max.iter.cluster = 20,
  harmony.epsilon.cluster = 1e-05,
  harmony.epsilon.harmony = 1e-04,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{n.dims.use}{The number of PCA dimensions used for the input for harmony. Default 30}

\item{fGSEA.minSize}{The cutoff minimum number of genes in each gene set. Gene set that contains the number of genes lower than this number will be excluded. Default 10}

\item{fGSEA.maxSize}{The cutoff maxiumum number of genes in each gene set. Gene set that contains the number of genes higher than this number will be excluded. Default 500}

\item{fGSEA.eps}{The eps paramenter for fgsea, this parameter sets the boundary for calculating the p value. Default 1e-10}

\item{hcl.height.cutoff}{The cutree cutoff value for hierarchical clustering. Default 0.25}

\item{covariates}{The unwanted source of variations (e.g. batch, sample_id, etc).}

\item{harmony.sigma}{The harmony sigma parameter. Default 0.1}

\item{harmony.tau}{The harmony tau parameter. Default 0}

\item{harmony.block.size}{The harmony block.size parameter. Default 0.05}

\item{harmony.max.iter}{The harmony max iteration parameter. Default 10}

\item{harmony.max.iter.cluster}{The harmony max iteration cluster. Default 200}

\item{harmony.epsilon.cluster}{The harmony epsilon.cluster parameter. Default 1e-05}

\item{harmony.epsilon.harmony}{The harmony epsilon parameter. Default 1e-04}

\item{n.seed}{The set seed number.}
}
\description{
Run supervised harmony integration
}
