% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runTSNE}
\alias{runTSNE}
\title{TSNE analysis}
\usage{
runTSNE(
  object,
  dim_reduction_method = c("pca", "nnmf"),
  useIntegrativeEmbeddings = F,
  integrative_method = c("combat", "seurat", "harmony", "supervised_harmony"),
  n.dims.use = 50,
  n.dims.embed = 2,
  n.perplexity = 30,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{dim_reduction_method}{The dimensional reduction method name that specifies the embedding matrix.}

\item{useIntegrativeEmbeddings}{is logical, if TRUE the embedding matrix genereated from integrative analysis will be used.}

\item{integrative_method}{The name of an integrative method.}

\item{n.dims.use}{The number of dimensions as the input.}

\item{n.dims.embed}{The number of output dimension. Default 2}

\item{n.perplexity}{The TSNE perplexity. Default 30}

\item{n.seed}{The number of set seed.}
}
\description{
TSNE analysis
}
