% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{runUMAP}
\alias{runUMAP}
\title{UMAP analysis}
\usage{
runUMAP(
  object,
  dim_reduction_method = c("pca", "nnmf", "lsi"),
  useIntegrativeEmbeddings = FALSE,
  integrative_method = c("combat", "seurat", "harmony", "supervised_harmony", "liger"),
  umap_method = c("uwot"),
  n.dims.use = 50,
  n.neighbors = 30,
  n.seed = 1,
  uwot.metric = "cosine",
  uwot.n.epochs = NULL,
  uwot.learning.rate = 1,
  uwot.min.dist = 0.25,
  uwot.spread = 1,
  uwot.set.op.mix.ratio = 1,
  uwot.local.connectivity = 1L,
  uwot.repulsion.strength = 1,
  uwot.negative.sample.rate = 5,
  uwot.a = NULL,
  uwot.b = NULL,
  uwot.metric.kwds = NULL,
  uwot.angular.rp.forest = FALSE,
  uwot.verbose = TRUE,
  uwot.save.model = FALSE,
  uwot.save.file = "uwot.out.rdata"
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{dim_reduction_method}{Dimensional reduction method name that specifies the embedding matrix.}

\item{useIntegrativeEmbeddings}{is logical, if TRUE the embedding matrix genereated from integrative analysis will be used.}

\item{integrative_method}{The name of an integrative method.}

\item{umap_method}{The seleced UMAP analysis method.}

\item{n.dims.use}{The number of dimensions as the input.}

\item{n.neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Default 30}

\item{n.seed}{The number of seed.}

\item{uwot.metric}{uwot's distance metric. Default 'cosine'.}

\item{uwot.spread}{uwot spread parameter. Default 1.0}

\item{uwot.set.op.mix.ratio}{uwot set.op.mix.ratio parameter. Default 1.0}

\item{uwot.local.connectivity}{uwot local connectivity parameter. Default 1L}

\item{uwot.repulsion.strength}{uwot repulsion.strength parameter. Default 1}

\item{uwot.negative.sample.rate}{uwot negative.sample.rate parameter. Default 5}

\item{uwot.a}{uwot a parameter.}

\item{uwot.b}{uwot b parameter.}

\item{uwot.metric.kwds}{uwot metric.kwds parameter.}

\item{uwot.angular.rp.forest}{is the logical of uwot angular.rp.forest.}

\item{uwot.verbose}{is logical for uwot verbose parameter.}

\item{uwot.save.model}{is the logical. If TRUE, the uwot model will be saved.}

\item{uwot.save.file}{The file name of uwot model to be saved.}
}
\description{
UMAP analysis
}
