% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellClasses.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,SingCellaR-method}
\alias{show,SingCellaR_Int-method}
\title{show method for SingCellaR}
\usage{
\S4method{show}{SingCellaR}(object)
}
\arguments{
\item{object}{A SingCellaR object}
}
\description{
show method for SingCellaR
}
