% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellClasses.R
\docType{class}
\name{SingCellaR}
\alias{SingCellaR}
\alias{SingCellaR-class}
\title{The SingCellaR Class
The SingCellaR object stores all of information required for the analyses.
The followings are slots which store information.}
\description{
The SingCellaR Class
The SingCellaR object stores all of information required for the analyses.
The followings are slots which store information.
}
\section{Slots}{

\describe{
\item{\code{dir_path_10x_matrix}}{is the path to the directory that contains 'barcodes.tsv','genes.tsv', and 'matrix.mtx' matrices generated by the 10x Genomics Cell Ranger software.}

\item{\code{sample_uniq_id}}{is the unique name of the sample.}

\item{\code{genes.info}}{contains the dataframe of gene information.}

\item{\code{meta.data}}{contains the dataframe of cells information.}

\item{\code{regressout_matrix}}{contain the normalised expression matrix (in log1p scale) that was regressed out the confounders.}

\item{\code{pca.result}}{contains the PCA analysis result from the IRLBA package}

\item{\code{nnmf.result}}{contains the NNMF analysis result.}

\item{\code{tsne.result}}{contains the TSNE analysis result.}

\item{\code{diffusionmap.result}}{contains the diffusionmap analysis result.}

\item{\code{umap.result}}{contains the UMAP analysis result.}

\item{\code{lsi.result}}{contans the LSI analysis result.}

\item{\code{knn_graph.graph}}{contains the igraph object.}

\item{\code{knn_graph.layout}}{contains the graph layout.}

\item{\code{knn_graph.kmeans.cluster}}{contains the k-means clustering results derived from the 3D KNN graph layout.}

\item{\code{igraph.graph}}{contains the igraph object.}

\item{\code{fa2_graph.layout}}{contains the graph layout generated by the ForceAtlas2 software.}

\item{\code{sc.clusters}}{contains the cluster information.}

\item{\code{marker.genes}}{contains the list of identified marker genes per each cluster.}
}}

