% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellClasses.R
\docType{class}
\name{SingCellaR_Int}
\alias{SingCellaR_Int}
\alias{SingCellaR_Int-class}
\title{SingCellaR_Int Class is the extended SingCellaR class
SingCellaR_Int object stores all of information required for multiple samples integration analyses.
The followings are slots which store information.}
\description{
SingCellaR_Int Class is the extended SingCellaR class
SingCellaR_Int object stores all of information required for multiple samples integration analyses.
The followings are slots which store information.
}
\section{Slots}{

\describe{
\item{\code{dir_path_SingCellR_object_files}}{contains the directory name that contains SingCellaR objects from multiple individual sample.}

\item{\code{SingCellR_object_files}}{contains the vector of individual object names.}

\item{\code{Variable.genes}}{contains the list of identified variable genes.}

\item{\code{SingCellaR.individual.clusters}}{contains the list of identified clusters information per individual sample.}

\item{\code{GSEA.aligned.clusters}}{contains the list of GSEA results per indentified clusters in individual sample.}

\item{\code{Scanorama.integrative.matrix}}{contains Scannorama integrative matrix.}

\item{\code{Harmony.embeddings}}{contains Harmony integrative embeddings.}

\item{\code{SupervisedHarmony.embeddings}}{contains SupervisedHarmony integrative embeddings.}

\item{\code{Liger.embeddings}}{contains Liger integrative embeddings.}

\item{\code{Seurat.embeddings}}{contains Seurat integrative embeddings.}

\item{\code{Combat.embeddings}}{contains Combat integrative embeddings.}
}}

