% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{sub_clusters}
\alias{sub_clusters}
\title{Extracting selected clusters.}
\usage{
sub_clusters(
  object,
  cluster.type = c("louvain", "walktrap", "kmeans", "merged_louvain",
    "merged_walktrap", "merged_kmeans"),
  cluster_ids = c()
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{cluster.type}{The clustering method name.}

\item{cluster_ids}{The vector of cluster ids that will be removed e.g. c('cl1:cl2','cl9:cl10').}
}
\value{
new SingCellaR object with selected clusters.
}
\description{
Extracting selected clusters.
}
