% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{subsample_cells}
\alias{subsample_cells}
\title{Subsample cells from the whole data set}
\usage{
subsample_cells(object, n.subsample = 10000, n.seed = 123)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{n.subsample}{The number of subsample cells to be selected.}

\item{n.seed}{The seed number for the subsampling. This number is used for reproducing the same subsampling results.}
}
\description{
Subsample cells from the whole data set
}
