% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellClasses.R
\docType{class}
\name{TargetSeq}
\alias{TargetSeq}
\alias{TargetSeq-class}
\title{The TargetSeq Class is the extended SingCellaR class
The TargetSeq object stores all of information required for TARGET-Seq analyses.
The followings are slots which store information.}
\description{
The TargetSeq Class is the extended SingCellaR class
The TargetSeq object stores all of information required for TARGET-Seq analyses.
The followings are slots which store information.
}
\section{Slots}{

\describe{
\item{\code{GenesExpressionMatrixFile}}{contains the file name of gene expression}

\item{\code{CellsMetaDataFile}}{contains the dataframe of TARGET-seq cell information.}
}}

