% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{TargetSeq_filter_cells_and_genes}
\alias{TargetSeq_filter_cells_and_genes}
\title{TargetSeq processing for filtering out cells and genes}
\usage{
TargetSeq_filter_cells_and_genes(
  object,
  min_Reads = 2000,
  min_detected_genes = 500,
  min_percent_mito = 0,
  max_percent_mito = 10,
  min_percent_ERCC = 0,
  max_percent_ERCC = 50,
  genes_with_expressing_cells = 10
)
}
\arguments{
\item{object}{TargetSeq object.}

\item{min_Reads}{The minimum number of reads cutoff.}

\item{min_detected_genes}{The minimum number of genes detected per cell.}

\item{min_percent_mito}{The minimum percent of reads from mitocondrial genes.}

\item{max_percent_mito}{The maximum percent of reads from mitocondrial genes.}

\item{min_percent_ERCC}{The minimum percent of reads from ERCC.}

\item{max_percent_ERCC}{The maximum percent of reads from ERCC.}

\item{genes_with_expressing_cells}{The cutoff for genes expressing in expressing cells.}
}
\description{
TargetSeq processing for filtering out cells and genes
}
