% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{TargetSeq_load_cell_metadata}
\alias{TargetSeq_load_cell_metadata}
\title{Loading cell metdata from TARGET-Seq analysis}
\usage{
TargetSeq_load_cell_metadata(
  object,
  sep = "\\t",
  a_column_of_the_cell_names = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{sep}{The field separator character.}

\item{a_column_of_the_cell_names}{The column that contains cell names.}
}
\description{
Loading cell metdata from TARGET-Seq analysis
}
