% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Build_AUCell_Rankings}
\alias{Build_AUCell_Rankings}
\title{Building AUCell gene rankings}
\usage{
Build_AUCell_Rankings(
  object,
  AUCell_buildRankings.file = "",
  nCores = NULL,
  plotStats = FALSE,
  splitByBlocks = FALSE
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{AUCell_buildRankings.file}{The output file of AUCell rankings.}

\item{nCores}{The number of cores for running AUCell. Default 1}

\item{plotStats}{is logical. If TRUE, the AUCell stats will be displayed.}

\item{splitByBlocks}{Whether to split the matrix by blocks in the ranking calculation. Allows using multiple cores. FALSE by default. Required for using sparse matrices.}
}
\description{
Building AUCell gene rankings
}
