% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{preprocess_integration}
\alias{preprocess_integration}
\title{Preprocessing for data integration}
\usage{
preprocess_integration(object, mitochondiral_genes_start_with = "MT-")
}
\arguments{
\item{object}{The SingCellaR object.}

\item{mitochondiral_genes_start_with}{The unique prefix alphabets for mitocondrial genes such as 'MT-' for human and 'mt-' for mouse genes.}
}
\description{
Preprocessing for data integration
}
