% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellRNASeq.R
\name{Run_fGSEA_for_multiple_comparisons}
\alias{Run_fGSEA_for_multiple_comparisons}
\title{Run GSEA analysis for multiple comparisons}
\usage{
Run_fGSEA_for_multiple_comparisons(
  GSEAPrerankedGenes_list,
  gmt.file = "",
  minSize = 5,
  maxSize = 2500,
  nperm = 10000,
  n.seed = 1
)
}
\arguments{
\item{GSEAPrerankedGenes_list}{The list of preranked genes.}

\item{gmt.file}{The GMT file name.}

\item{minSize}{The cutoff minimum number of genes in each gene set. Gene set that contains the number of genes lower than this number will be excluded. Default 5}

\item{maxSize}{The cutoff maxiumum number of genes in each gene set. Gene set that contains the number of genes higher than this number will be excluded. Default 2500}

\item{nperm}{The number of fGSEA permutation. Default 10000}

\item{n.seed}{The set seed number.}

\item{object}{The SingCellaR object.}
}
\value{
a dataframe of GSEA analysis results.
}
\description{
Run GSEA analysis for multiple comparisons
}
