% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCell_Integration.R
\name{runHarmony}
\alias{runHarmony}
\title{Run harmony integration}
\usage{
runHarmony(
  object,
  n.dims.use = 30,
  covariates = c("data_set"),
  harmony.theta = NULL,
  harmony.sigma = 0.1,
  harmony.nclust = NULL,
  harmony.tau = 0,
  harmony.block.size = 0.05,
  harmony.max.iter = 10,
  harmony.max.iter.cluster = 200,
  harmony.epsilon.cluster = 1e-05,
  harmony.epsilon.harmony = 1e-04,
  n.seed = 1
)
}
\arguments{
\item{object}{The SingCellaR object.}

\item{n.dims.use}{The number of PCA dimensions used for the input for harmony.}

\item{covariates}{The unwanted source of variations (e.g. batch, sample_id, etc).}

\item{harmony.theta}{The harmony theta parameter. Deafult NULL. This is the diversity clustering penalty parameter. Specify for each variable in vars_use Default theta=2. theta=0 does not encourage any diversity. Larger values of theta result in more diverse clusters.}

\item{harmony.sigma}{The harmony sigma parameter. Default 0.1. This parameter is the width of soft kmeans clusters. Sigma scales the distance from a cell to cluster centroids. Larger values of sigma result in cells assigned to more clusters. Smaller values of sigma make soft kmeans cluster approach hard clustering.}

\item{harmony.nclust}{The harmony nclust parameter. Default NULL}

\item{harmony.tau}{The harmony tau parameter. Default 0. The parameter is for the protection against overclustering small datasets with large ones. tau is the expected number of cells per cluster.}

\item{harmony.block.size}{The harmony block.size parameter. Default 0.05. What proportion of cells to update during clustering. Between 0 to 1. Larger values may be faster but less accurate}

\item{harmony.max.iter}{The harmony max iteration parameter. Default 10. Maximum number of rounds to run Harmony. One round of Harmony involves one clustering and one correction step.}

\item{harmony.max.iter.cluster}{The harmony max iteration cluster. Default 200. Maximum number of rounds to run clustering at each round of Harmony.}

\item{harmony.epsilon.cluster}{The harmony epsilon.cluster parameter. Default 1e-05. Convergence tolerance for clustering round of Harmony. Set to -Inf to never stop early.}

\item{harmony.epsilon.harmony}{The harmony epsilon parameter. Default 1e-04. Convergence tolerance for Harmony. Set to -Inf to never stop early.}

\item{n.seed}{The set seed number.}
}
\description{
Run harmony integration
}
